refactored the cache out of the component and added it as a project service
