+++ src/com/haskforce/cabal/highlighting/CabalAnnotator.scala
@@ -1,10 +1,10 @@
 package com.haskforce.cabal.highlighting
 
+import com.haskforce.cabal.lang.psi.CabalTypes._
+import com.haskforce.cabal.lang.psi._
 import com.intellij.lang.annotation.{AnnotationHolder, Annotator, HighlightSeverity}
 import com.intellij.openapi.editor.colors.{EditorColorsManager, TextAttributesKey}
 import com.intellij.psi.PsiElement
-import com.haskforce.cabal.lang.psi._
-import com.haskforce.cabal.lang.psi.CabalTypes._
 
 class CabalAnnotator extends Annotator {
 
@@ -18,8 +18,9 @@ class CabalAnnotator extends Annotator {
     }
   }
 
-  private def setHighlighting(element: PsiElement, holder: AnnotationHolder, key: TextAttributesKey) {
+  private def setHighlighting(element: PsiElement, holder: AnnotationHolder, key: TextAttributesKey): Unit = {
     holder.newSilentAnnotation(HighlightSeverity.INFORMATION)
       .enforcedTextAttributes(EditorColorsManager.getInstance.getGlobalScheme.getAttributes(key))
+    ()
   }
 }
